+++ reductstore/src/storage/query/condition/operators/string.rs
@@ -1,0 +1,10 @@
+// Copyright 2024 ReductSoftware UG
+// Licensed under the Business Source License 1.1
+
+mod contains;
+mod ends_with;
+mod starts_with;
+
+pub(crate) use contains::Contains;
+pub(crate) use ends_with::EndsWith;
+pub(crate) use starts_with::StartsWith;