+++ reductstore/src/storage/query/condition/value/string.rs
@@ -1,0 +1,6 @@
+// Copyright 2025 ReductSoftware UG
+// Licensed under the Business Source License 1.1
+
+pub(crate) mod contains;
+pub(crate) mod ends_with;
+pub(crate) mod starts_with;